+++ addon/services/groups-service.js
@@ -79,8 +79,8 @@ export default Service.extend(serviceMixin, {
    */
   create (group, portalOpts) {
     let clonedGroup = JSON.parse(JSON.stringify(group));
-    // JSON parse / json stringify utterly mucks up the thumbnail object so...
-    if (group.thumbnail) {
+    // cloning can't handle Blobs, so we pass the thumbnail property forward
+    if (typeof Blob !== "undefined" && group.thumbnail instanceof Blob) {
       clonedGroup.thumbnail = group.thumbnail;
     }
 
@@ -99,8 +99,8 @@ export default Service.extend(serviceMixin, {
    */
   update (group, portalOpts) {
     let clonedGroup = JSON.parse(JSON.stringify(group));
-    // JSON parse / json stringify utterly mucks up the thumbnail object so...
-    if (group.thumbnail) {
+    // cloning can't handle Blobs, so we pass the thumbnail property forward
+    if (typeof Blob !== "undefined" && group.thumbnail instanceof Blob) {
       clonedGroup.thumbnail = group.thumbnail;
     }